requirements.txt
discord.py==2.0.1
flask==2.0.3
googletrans==4.0.0-rc1
python-dotenv==0.19.2
Werkzeug==2.0.3